describe camada mais detalhado
